claasses location have been changed.

+++ academichub/src/main/java/com/yazlab/academichub/entities/ArticleType.java
@@ -20,5 +20,5 @@ public class ArticleType {
 
     private int articleType;
 
-    private int coefficient;
+    private double coefficient;
 }
